[wip] Implement fastapi router and deps for auth endpoints
