+++ templates/google_page.html
@@ -59,7 +59,7 @@
                 <div class="w-full py-4 px-6">
                     <div x-show="currentTab === 0">
                         <div class="flex my-6 justify-center">
-                            <form action="/login-google" method="GET">
+                            <form action="/auth/login-google" method="GET">
                                 <button class="gsi-material-button border border-gray-400 rounded-lg p-2" type='submit'>
                                     <div class="gsi-material-button-content-wrapper flex items-center space-x-2">
                                         <div class="gsi-material-button-icon">
